Modify randomGearLoweringChance
Added second roll for randomGearLoweringChance so item quality can randomly decrease up to two quality levels (e.g when config is set to epic quality there is a chance the piece will be rare, and a second roll to determine if it will decrease again to uncommon). Will not decrease below Common quality regardless of settings.